requirements.txt
Flask==3.0.0
readability-lxml==0.8.1
beautifulsoup4==4.12.2
lxml==4.9.3
Werkzeug==3.0.1
